backend/ReMu/requirements.txt
absl-py==0.10.0
aiohttp==3.6.2
appdirs==1.4.4
asgiref==3.2.10
astunparse==1.6.3
async-timeout==3.0.1
attrs==20.2.0
audioread==2.1.8
Automat==20.2.0
beautifulsoup4==4.9.1
browser-cookie3==0.11.2
brukeropusreader==1.3.4
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
coloredlogs==14.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
crepe==0.0.11
cryptography==3.1
cssselect==1.1.0
cycler==0.10.0
decorator==4.4.2
Django==3.0.5
django-cors-headers==3.5.0
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
djongo==1.3.3
drf-yasg==1.17.1
ffmpeg==1.4
ffmpeg-python==0.2.0
future==0.18.2
gast==0.3.3
google-auth==1.22.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.32.0
h11==0.9.0
h5py==2.10.0
hmmlearn==0.2.4
httpcore==0.10.2
httpx==0.14.3
humanfriendly==8.2
hyperlink==20.0.1
idna==2.10
imageio==2.9.0
incremental==17.5.0
inflection==0.5.1
itemadapter==0.1.0
itemloaders==1.0.3
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.16.0
Keras-Preprocessing==1.1.2
keyring==21.4.0
kiwisolver==1.2.0
librosa==0.8.0
llvmlite==0.34.0
lxml==4.5.2
lyricwikia==0.1.11
lz4==3.1.0
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.3.2
multidict==4.7.6
mutagen==1.45.1
numba==0.51.2
numexpr==2.7.1
numpy==1.18.5
oauthlib==3.1.0
openapi-codec==1.3.2
opt-einsum==3.3.0
packaging==20.4
pandas==1.1.1
parsel==1.6.0
pathlib==1.0.1
pbkdf2==1.3
Pillow==7.2.0
pooch==1.2.0
Protego==0.1.16
protobuf==3.13.0
psycopg2==2.8.6
pyaes==1.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.9.8
PyDispatcher==2.0.5
PyHamcrest==2.0.2
PyJWT==1.7.1
pymongo==3.11.0
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyreadline==2.1
python-dateutil==2.8.1
pytube3==9.6.4
pytz==2020.1
pywin32-ctypes==0.2.0
PyYAML==5.3.1
queuelib==1.5.0
requests==2.24.0
requests-oauthlib==1.3.0
resampy==0.2.2
rfc3986==1.4.0
rsa==4.6
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
scikit-learn==0.23.2
scipy==1.5.2
Scrapy==2.3.0
selenium==3.141.0
service-identity==18.1.0
simplejson==3.17.2
six==1.15.0
sklearn==0.0
sniffio==1.1.0
SoundFile==0.10.3.post1
soupsieve==2.0.1
spotdl==2.2.2
spotipy==2.14.0
sqlparse==0.2.4
tables==3.6.1
tekore==3.1.0
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
threadpoolctl==2.1.0
tinytag==1.4.0
tqdm==4.48.2
Twisted==20.3.0
typing-extensions==3.7.4.3
unicode-slugify==0.1.3
Unidecode==1.1.1
uritemplate==3.0.1
urllib3==1.25.10
w3lib==1.22.0
websocket-client==0.57.0
websockets==8.1
Werkzeug==1.0.1
wincertstore==0.2
wrapt==1.12.1
yarl==1.6.0
youtube-dl==2020.9.6
zope.interface==5.1.0
